// src/app/models/article.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
/* eslnumber-disable @typescript-eslnumber/naming-convention */
export class ARTICLE {
  	ID?: number;
	NATURE?: number;
	RACCOURCI?: number;
	ARTICLE?: string;
	ORIGINE?: string;
	COMPOSE?: string;
	REFERENCE?: string;
	CoMatiere?: number;
	RemiseAccorde?: number;
	CoFabrication?: number;
	DroitConsommation?: number;
	PrixHT?: number;
	FraisChargement?: number;
	FraisTransport?: number;
	AutreFraisPerCent?: number;
	AutreFrais?: number;
	MargeSecurite?: number;
	TVAachat?: number;
	AutreTaxe?: number;
	PrixRevient?: number;
	PrixRevientTTC?: number;
	Remise1?: number;
	Remise2?: number;
	Remise3?: number;
	Remise4?: number;
	Remise5?: number;
	Remise6?: number;
	UniteAchat?: string;
	CollisageAchat?: number;
	PoidsAchat?: number;
	CalculPrixVente?: number;
	Marge?: number;
	PrixVHT?: number;
	Fodec?: number;
	GICA?: number;
	TVA?: number;
	PrixVTTC?: number;
	Unite?: string;
	CollisageVente?: number;
	Poids?: number;
	FacteurStock?: number;
	Homologue?: number;
	SM?: number;
	PVPon?: number;
	PVP?: number;
	emballage?: number;
	RemiseMax?: number;
	PMP?: number;
	DPA?: number;
	DATEENTREE?: string;
	StockPMP?: number;
	StockDPA?: number;
	DATESORTIE?: Date;
	StockMin?: number;
	StockMax?: number;
	StockAlerte?: number;
	StockCmd?: number;
	StockReel?: number;
	StockAchat?: number;
	STOCK_INITIAL?: number;
	STOCK_ENTREE?: number;
	STOCK_CASSE?: number;
	STOCK_SORTIE?: number;
	STOCK?: number;
	REGION?: string;
	GC?: number;
	FLAG?: number;
	ANNEE?: string;
	FAMILLE?: number;
	SOUSFAMILLE?: number;
	CATEGORIE?: number;
	CLASSE?: number;
	FILTRE1?: number;
	FILTRE2?: number;
	FILTRE3?: number;
	FILTRE4?: number;
	FILTRE5?: number;
	FILTRE6?: number;
	CIN_FRN?: string;
	FRN?: string;
	QteCmd?: number;
	DELAI?: number;
	PromotionOn?: number;
	PromotionDebut?: Date;
	PromotionFin?: Date;
	PromotionCalcul?: number;
	PromotionMarge?: number;
	PromotionPrixHT?: number;
	PromotionPrixTTC?: number;
	T1_TarifOn?: number;
	T1_QteDebut?: number;
	T1_QteFin?: number;
	T1_Calcul?: number;
	T1_Marge?: number;
	T1_PrixHT?: number;
	T1_PrixTTC?: number;
	T2_TarifOn?: number;
	T2_QteDebut?: number;
	T2_QteFin?: number;
	T2_Calcul?: number;
	T2_Marge?: number;
	T2_PrixHT?: number;
	T2_PrixTTC?: number;
	T3_TarifOn?: number;
	T3_QteDebut?: number;
	T3_QteFin?: number;
	T3_Calcul?: number;
	T3_Marge?: number;
	T3_PrixHT?: number;
	T3_PrixTTC?: number;
	DATECREATION?: string;
	UTILISATEUR1?: string;
	DATEMODIF?: string;
	UTILISATEUR2?: string;
	MOLDELE?: string;
	MARQUE?: string;
	pFIDELITE?: number;
	PrixInitial?: number;
	E2012?: number;
	S2012?: number;
	I2012?: number;
	R2012?: number;
	ROTATION?: number;
	Q1?: number;
	Q2?: number;
	pQ1?: number;
	pQ2?: number;
	A01?: string;
	A02?: string;
	A03?: string;
	A04?: string;
	A05?: string;
	A06?: string;
	A07?: string;
	A08?: string;
	A09?: string;
	A10?: string;
	A11?: string;
	A12?: string;
	A13?: string;
	I2015?: number;
	I2014?: number;
	S2015?: number;
	E2015?: number;
	DELTA?: number;
	M2015?: number;
	ECART?: number;
	STK_INFO?: number;
	ACHAT_TTC?: number;
	txtFAMILLE?: string;
	codeFamille?: string;
	CAMION?: string;
}