// src/app/models/parametreCamionPDA.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
export class ParametreCamionPDA {
	Id?: number;
	CodeCamion?: string;
	CodePDA?: string;
	MatriculeCamion?: string;
	Initialiser?: number;
	Date?: Date;
	DernierNumBL?: number;
	DernierNumBE?: number;
	SetCodeClient?: number;
	Version?: number;
	DateMiseAJour?: Date;
	Facture?: number;
	Bonsortie?: number;
	Prefix?: string;
}