// src/app/models/bl_ligne.model.ts
/* eslint-disable @typescript-eslint/naming-convention */
export class BLLigne{
	IdBLLigne?: number;
	CodeArticle?: string;
	IdCamionPDA?: string;
	Quantiter?: number;
	PrixHT?: number;
	TVA?: number;
	Remise?: number;
	PrixTTC?: number;
	Unite?: string;
	BLNum?: string;
	BLType?: string;
	IdPromotion?: number;
	Point?: number;
}